use i32

+++ src/host/handler.rs
@@ -4,7 +4,7 @@ use crate::host::memory;
 mod http_handler;
 
 pub fn log(level: i32, message: &[u8]) {
-    unsafe { http_handler::log(level, message.as_ptr(), message.len() as u32) };
+    unsafe { http_handler::log(level, message.as_ptr(), message.len() as i32) };
 }
 
 pub fn log_enabled(level: i32) -> bool {
@@ -25,59 +25,59 @@ pub fn get_config(buffer: &Buffer) -> Option<String> {
     }
 }
 
-pub fn enable_feature(feature: u32) -> i32 {
+pub fn enable_feature(feature: i32) -> i32 {
     unsafe { http_handler::enable_features(feature) }
 }
 
-pub fn header_values(buffer: &Buffer, kind: u32, name: &[u8]) -> Vec<Vec<u8>> {
+pub fn header_values(buffer: &Buffer, kind: i32, name: &[u8]) -> Vec<Vec<u8>> {
     match unsafe {
         http_handler::get_header_values(
             kind,
             name.as_ptr(),
-            name.len() as u32,
+            name.len() as i32,
             buffer.data.as_ptr(),
             buffer.size,
         )
     } {
-        count_len if (count_len as u32) <= (buffer.size as u32) => {
+        count_len if (count_len as i32) <= buffer.size => {
             memory::handle_values(buffer.data.as_slice(), count_len)
         }
         _ => todo!(),
     }
 }
-pub fn header_names(buffer: &Buffer, kind: u32) -> Vec<Vec<u8>> {
+pub fn header_names(buffer: &Buffer, kind: i32) -> Vec<Vec<u8>> {
     match unsafe { http_handler::get_header_names(kind, buffer.data.as_ptr(), buffer.size) } {
-        count_len if (count_len as u32) <= (buffer.size as u32) => {
+        count_len if (count_len as i32) <= buffer.size => {
             memory::handle_values(buffer.data.as_slice(), count_len)
         }
         _ => todo!(),
     }
 }
 
-pub fn remove_header(kind: u32, name: &[u8]) {
-    unsafe { http_handler::remove_header(kind, name.as_ptr(), name.len() as u32) }
+pub fn remove_header(kind: i32, name: &[u8]) {
+    unsafe { http_handler::remove_header(kind, name.as_ptr(), name.len() as i32) }
 }
 
-pub fn set_header(kind: u32, name: &[u8], value: &[u8]) {
+pub fn set_header(kind: i32, name: &[u8], value: &[u8]) {
     unsafe {
         http_handler::set_header_value(
             kind,
             name.as_ptr(),
-            name.len() as u32,
+            name.len() as i32,
             value.as_ptr(),
-            value.len() as u32,
+            value.len() as i32,
         )
     };
 }
 
-pub fn add_header_value(kind: u32, name: &[u8], value: &[u8]) {
+pub fn add_header_value(kind: i32, name: &[u8], value: &[u8]) {
     unsafe {
         http_handler::add_header_value(
             kind,
             name.as_ptr(),
-            name.len() as u32,
+            name.len() as i32,
             value.as_ptr(),
-            value.len() as u32,
+            value.len() as i32,
         )
     };
 }
@@ -97,21 +97,21 @@ pub fn method(buffer: &Buffer) -> Option<Vec<u8>> {
 }
 
 pub fn set_method(method: &[u8]) {
-    unsafe { http_handler::set_method(method.as_ptr(), method.len() as u32) };
+    unsafe { http_handler::set_method(method.as_ptr(), method.len() as i32) };
 }
 
 pub fn set_uri(uri: &[u8]) {
-    unsafe { http_handler::set_uri(uri.as_ptr(), uri.len() as u32) };
+    unsafe { http_handler::set_uri(uri.as_ptr(), uri.len() as i32) };
 }
 
 pub fn version(buffer: &Buffer) -> Option<Vec<u8>> {
-    match unsafe { http_handler::get_protocol_version(buffer.data.as_ptr(), buffer.size as u32) } {
+    match unsafe { http_handler::get_protocol_version(buffer.data.as_ptr(), buffer.size) } {
         size if size <= buffer.size => memory::to_bytes(buffer.data.as_slice(), size),
         _ => None,
     }
 }
 pub fn uri(buffer: &Buffer) -> Option<Vec<u8>> {
-    match unsafe { http_handler::get_uri(buffer.data.as_ptr(), buffer.size as u32) } {
+    match unsafe { http_handler::get_uri(buffer.data.as_ptr(), buffer.size) } {
         size if size <= buffer.size => memory::to_bytes(buffer.data.as_slice(), size),
         _ => None,
     }
@@ -125,13 +125,13 @@ pub fn set_status_code(code: i32) {
     unsafe { http_handler::set_status_code(code) }
 }
 
-pub fn body(buffer: &Buffer, kind: u32) -> Option<Vec<u8>> {
+pub fn body(buffer: &Buffer, kind: i32) -> Option<Vec<u8>> {
     let mut eof = false;
     let mut size;
     let mut out = Vec::new();
     while !eof {
         (eof, size) = memory::eof_size(unsafe {
-            http_handler::read_body(kind, buffer.data.as_ptr(), buffer.size as u32)
+            http_handler::read_body(kind, buffer.data.as_ptr(), buffer.size)
         });
         if let Some(vec) = memory::to_bytes(buffer.data.as_slice(), size) {
             out.push(vec)
@@ -144,9 +144,9 @@ pub fn body(buffer: &Buffer, kind: u32) -> Option<Vec<u8>> {
     }
 }
 
-pub fn write_body(kind: u32, body: &str) {
+pub fn write_body(kind: i32, body: &str) {
     unsafe {
-        http_handler::write_body(kind, body.as_ptr(), body.len() as u32);
+        http_handler::write_body(kind, body.as_ptr(), body.len() as i32);
     }
 }
